+++ Frontend/src/app/Components/Utils/navbar/navbar.component.ts
@@ -1,13 +1,37 @@
-import { Component } from '@angular/core';
-import { RouterLink } from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import { Router, RouterLink } from '@angular/router';
+import { AuthService } from '../../../Services/Authentication/auth.service';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-navbar',
   standalone: true,
-  imports: [RouterLink],
+  imports: [RouterLink, CommonModule],
   templateUrl: './navbar.component.html',
-  styleUrl: './navbar.component.css'
+  styleUrls: ['./navbar.component.css']
 })
-export class NavbarComponent {
+export class NavbarComponent implements OnInit {
+  isAuthenticated: boolean = false;
+  user: any;
 
+  constructor(private authService: AuthService, private router: Router) {}
+
+  ngOnInit() {
+    this.checkAuthentication();
+  }
+
+  checkAuthentication() {
+    // Check if the user is authenticated
+    this.isAuthenticated = this.authService.isAuthenticated();
+    if (this.isAuthenticated) {
+      this.user = this.authService.currentUserValue; // Get the logged-in user
+    }
+  }
+
+  logout() {
+    // Call the AuthService to log out the user
+    this.authService.logout();
+    window.location.reload();
+    this.router.navigate(['/home']);  // Redirect to home after logout
+  }
 }